index.md: simplify 3 May 2013
Based on a diff written by @lorenzocutelle. I had to play jigsaw
to apply it. Apparently, it is not a big deal to send bathes of
changes because the web interface is not very good in generating
a diff (so changes from others are incorporated).

We'd better send pull requests for small changes; an alternative
it is to use the command line, but I guess this is to complex.