Move .htpasswd to studenti
Because I would like to use a private git repository to keep
track of the data and because I would like to track the passwords
as well, I am moving the password file to studenti.

The mechanism to track the content of the studenti directory
using git will come next.