Serve jquery.md5.min.js locally
Chromium refused to serve jquery.md5.min.js from raw.github.com,
because the MIME type was not compatible with javascript and (this
is my guess) because it was an external site.

Hence, the mechanism to login once was not working.

Fix by providing such javascript file locally; a more comprehensive
solution (including update of the script) may follow.