Accept URIs containing & characters
This diff tweaks the regular expressions in backend.js to also accept
URIs contaning & characters.

Prompted by complaints from students on the mailing list.

While at it: tweak the output of the private.tpl.html page and make sure
that, if the personal page cannot be updated, an error will be returned to
the user.